+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## vNext
+
+- Add note to exception if Tailscale is not installed properly
+
 ## 0.7.4
 
 - Refactor exception logging arguments